implementing async await in UIKit demo
