+++ UIKitDemo/UIKitDemo/PaymentViewController.swift
@@ -10,7 +10,7 @@ import SwiftUI
 import MoyasarSdk
 import PassKit
 
-let handler = ApplePayPaymentHandler()
+
 /// No `allowedNetworks` set, defaults to all networks
 /// allowedNetworks: [.visa, .mastercard] // Only accept Visa and Mastercard
 ///  Amount in the smallest currency unit.
@@ -41,6 +41,8 @@ let token = ApiTokenRequest(
 
 class PaymentViewController: UIViewController {
     
+    let handler = ApplePayPaymentHandler(paymentRequest: paymentRequest)
+
     override func viewDidLoad() {
         view.backgroundColor = .white
         
@@ -106,70 +108,5 @@ class PaymentViewController: UIViewController {
     
     func handleApplePayPressed(action: UIAction) {
         handler.present()
-    }
-}
-
-class ApplePayPaymentHandler: NSObject, PKPaymentAuthorizationControllerDelegate {
-    var applePayService = ApplePayService(apiKey: "pk_test_vcFUHJDBwiyRu4Bd3hFuPpTnRPY4gp2ssYdNJMY3")
-    var controller: PKPaymentAuthorizationController?
-    var items = [PKPaymentSummaryItem]()
-    var networks: [PKPaymentNetwork] = [
-        .amex,
-        .mada,
-        .masterCard,
-        .visa
-    ]
-    
-    override init() {}
-    
-    func present() {
-        items = [
-            PKPaymentSummaryItem(label: "Moyasar", amount: 1.00, type: .final)
-        ]
-        
-        let request = PKPaymentRequest()
-        
-        request.paymentSummaryItems = items
-        request.merchantIdentifier = "merchant.nuha.io.second"
-        request.countryCode = "SA"
-        request.currencyCode = "SAR"
-        request.supportedNetworks = networks
-        request.merchantCapabilities = [
-            .capability3DS,
-            .capabilityCredit,
-            .capabilityDebit
-        ]
-        
-        controller = PKPaymentAuthorizationController(paymentRequest: request)
-        controller?.delegate = self
-        controller?.present(completion: {(p: Bool) in
-            print("Presented: " + (p ? "Yes" : "No"))
-        })
-    }
-    
-    func paymentAuthorizationController(_ controller: PKPaymentAuthorizationController, didAuthorizePayment payment: PKPayment, completion: @escaping (PKPaymentAuthorizationStatus) -> Void) {
-        
-        do {
-            try applePayService.authorizePayment(request: paymentRequest, token: payment.token) {(result: ApiResult<ApiPayment>) in
-                switch (result) {
-                case .success(let payment):
-                    print("Got payment")
-                    print(payment.status)
-                    print(payment.id)
-                    break;
-                case .error(let error):
-                    print(error)
-                    break;
-                }
-            }
-        } catch {
-            print(error)
-        }
-        
-        completion(.success)
-    }
-    
-    func paymentAuthorizationControllerDidFinish(_ controller: PKPaymentAuthorizationController) {
-        controller.dismiss(completion: {})
     }
 }
